requirements.txt
torch
torchvision
numpy
scipy
pillow
matplotlib
pandas
jupyter